requirements.txt
requests~=2.25.1
youtube-dl==2021.2.10
discord~=1.0.1
python-dotenv~=0.15.0
bs4~=0.0.1
beautifulsoup4~=4.9.3
setuptools~=53.1.0
PyNaCl==1.4.0
imgurpython~=1.1.7
Cogs~=0.4.4
Pillow~=8.1.0
Wand~=0.6.6
faces~=0.4
Unidecode~=1.2.0
lxml
PornhubApi